+++ BlueberryEditor/src/Editor/Preview/MaterialPreview.cpp
@@ -43,7 +43,7 @@ namespace Blueberry
 			m_Camera->SetOrthographic(false);
 			m_Camera->SetAspectRatio(1.0f);
 			m_Camera->SetFieldOfView(15.0f);
-			m_Camera->SetPixelSize(Vector2(target->GetWidth(), target->GetHeight()));
+			m_Camera->SetPixelSize(Vector2(static_cast<float>(target->GetWidth()), static_cast<float>(target->GetHeight())));
 			m_Camera->SetCameraType(CameraType::Preview);
 
 			Entity* skyEntity = m_Scene->CreateEntity("Sky");
